fills out readme, tweaks kyc details in example
