+++ @stellar/typescript-wallet-sdk/examples/helpers/ask-question.ts
@@ -1,0 +1,15 @@
+import readline from "readline";
+
+export const askQuestion = (query) => {
+  const rl = readline.createInterface({
+    input: process.stdin,
+    output: process.stdout,
+  });
+
+  return new Promise<string>((resolve) =>
+    rl.question(query, (ans) => {
+      rl.close();
+      resolve(ans);
+    }),
+  );
+};
