+++ @stellar/typescript-wallet-sdk/package.json
@@ -69,6 +69,7 @@
     "build:web": "webpack --config webpack.config.js",
     "build:node": "webpack --env NODE=true --config webpack.config.js",
     "build": "run-p build:web build:node",
-    "example:sep24": "ts-node examples/sep24/sep24.ts"
+    "example:sep24": "ts-node examples/sep24/sep24.ts",
+    "example:sep12": "ts-node examples/sep12/sep12.ts"
   }
 }
